Rename package to plural form
