+++ server/src/main/java/com/hawktu/server/dtos/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.hawktu.server.dto.api.response;
+package com.hawktu.server.dtos.response;
 
 public class ErrorResponse {
     private String message;
