+++ server/src/main/java/com/hawktu/server/utils/JwtUtil.java
@@ -1,10 +1,10 @@
-package com.hawktu.server;
+package com.hawktu.server.utils;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.security.core.userdetails.UserDetails;
+// import io.jsonwebtoken.security.Keys;
+// import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 
 import javax.crypto.SecretKey;
